refac: move generateInitialsImage function to utils
